+++ module/action.sh
@@ -14,4 +14,4 @@ if [ -z "$MMRL" ] && [ ! -z "$MAGISKTMP" ]; then
 fi
 
 sh $MODDIR/autopif_ota.sh || true
-sh $MODDIR/autopif.sh -p
+sh $MODDIR/autopif.sh